added search functionality
added the ability to search via a method that takes in the given list of
data, a string for the data to be found, and the nameInHeader of where
the data will be found.

also broke the code to print the screen out of the main method and
created its own method for printing all data to the console.